Delete poti sa faci doar pe cursurile pe ai caror teacheri ai dreptul. In plus: ROLE_ADMIN cand definesti permisiunile userului logat.

+++ spring-one/src/main/java/victor/training/spring/web/controller/TrainingController.java
@@ -7,8 +7,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import victor.training.spring.web.controller.dto.TrainingDto;
+import victor.training.spring.web.domain.Training;
+import victor.training.spring.web.repo.TrainingRepo;
+import victor.training.spring.web.security.SecurityUser;
 import victor.training.spring.web.service.TrainingService;
 
 @RestController
@@ -44,8 +49,24 @@ public class TrainingController {
 	@PreAuthorize("hasRole('ADMIN')")
 	@DeleteMapping("{id}")
 	public void deleteTrainingById(@PathVariable Long id) {
+		Training training = trainingRepo.findById(id).get();
+
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		SecurityUser principal = (SecurityUser) authentication.getPrincipal();
+
+		if (!principal.getManagedTeacherIds().contains(training.getTeacher().getId())) {
+			throw  new RuntimeException("ia mana!");
+		}
+
+
+
 		trainingService.deleteById(id);
 	}
+
+	@Autowired
+	private TrainingRepo trainingRepo;
+
+
 
 	@PostMapping
 	public void createTraining(@RequestBody TrainingDto dto) throws ParseException {
